# Optimize Hash and isRepetition function