--- a/bench/src/main/scala/benchmarks/HashBench.scala
+++ b/bench/src/main/scala/benchmarks/HashBench.scala
@@ -37,3 +37,17 @@ class HashBench:
       Blackhole.consumeCPU(Work)
       Hash(x)
     bh.consume(result)
+
+  @Benchmark
+  def repetition5(bh: Blackhole) =
+    var result = situations.map: x =>
+      Blackhole.consumeCPU(Work)
+      x.board.history.fivefoldRepetition
+    bh.consume(result)
+
+  @Benchmark
+  def repetition3(bh: Blackhole) =
+    var result = situations.map: x =>
+      Blackhole.consumeCPU(Work)
+      x.board.history.threefoldRepetition
+    bh.consume(result)