--- a/src/main/scala/History.scala
+++ b/src/main/scala/History.scala
@@ -1,7 +1,6 @@
 package chess
 
 import format.Uci
-import cats.kernel.Monoid
 
 // Checks received by the respective side.
 case class CheckCount(white: Int = 0, black: Int = 0):
@@ -18,7 +17,7 @@ case class CheckCount(white: Int = 0, black: Int = 0):
 
 case class History(
     lastMove: Option[Uci] = None,
-    positionHashes: PositionHash = Monoid[PositionHash].empty,
+    positionHashes: PositionHash = PositionHash.empty,
     castles: Castles = Castles.init,
     checkCount: CheckCount = CheckCount(0, 0),
     unmovedRooks: UnmovedRooks,
@@ -27,22 +26,8 @@ case class History(
 
   def setHalfMoveClock(v: HalfMoveClock) = copy(halfMoveClock = v)
 
-  // private def isRepetition(times: Int) =
-  //   positionHashes.value.length > (times - 1) * 4 * Hash.size && {
-  //     // compare only hashes for positions with the same side to move
-  //     val positions = positionHashes.value.sliding(Hash.size, 2 * Hash.size).toList
-  //     positions.headOption match
-  //       case Some(Array(x, y, z)) =>
-  //         (positions count {
-  //           case Array(x2, y2, z2) => x == x2 && y == y2 && z == z2
-  //           case _                 => false
-  //         }) >= times
-  //       case _ => times <= 1
-  //   }
-
   inline def threefoldRepetition = positionHashes.isRepetition(3)
-
-  inline def fivefoldRepetition = positionHashes.isRepetition(5)
+  inline def fivefoldRepetition  = positionHashes.isRepetition(5)
 
   inline def canCastle(inline color: Color)                    = castles can color
   inline def canCastle(inline color: Color, inline side: Side) = castles.can(color, side)
@@ -76,7 +61,7 @@ object History:
       lastMove = lastMove flatMap Uci.apply,
       castles = Castles(castles),
       unmovedRooks = UnmovedRooks.corners,
-      positionHashes = Monoid[PositionHash].empty
+      positionHashes = PositionHash.empty
     )
 
   def castle(color: Color, kingSide: Boolean, queenSide: Boolean) =