Before

```
[info] HashBench.hashes       thrpt   45  746.230 ± 15.666  ops/s
[info] HashBench.repetition3  thrpt   45  709.714 ± 26.750  ops/s
[info] HashBench.repetition5  thrpt   45  808.189 ± 32.930  ops/s
```

After removing newtypes & change algorithm

```
[info] HashBench.hashes       thrpt   45    748.412 ±  19.580  ops/s
[info] HashBench.repetition3  thrpt   45  12099.058 ± 232.333  ops/s
[info] HashBench.repetition5  thrpt   45  12276.818 ± 187.127  ops/s
```

Removing Monoid

```
[info] HashBench.hashes       thrpt   45    758.422 ±  33.644  ops/s
[info] HashBench.repetition3  thrpt   45  12476.803 ± 114.103  ops/s
[info] HashBench.repetition5  thrpt   45  12865.016 ± 195.063  ops/s
```

[Full benchmark](https://gist.github.com/lenguyenthanh/9380c5422ca40efedb567c32de1f4793)